VisionAssistant_Pi/requirements.txt
torch>=2.1
transformers>=4.36
pillow
opencv-python
sox

